--- src/main/java/com/heeverse/config/SqlSessionFactoryConfiguration.java
@@ -1,41 +1,0 @@
-package com.heeverse.config;
-
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.SqlSessionTemplate;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.core.io.ClassPathResource;
-
-import javax.sql.DataSource;
-
-/**
- * @author jeongheekim
- * @date 10/28/23
- */
-@Configuration
-@MapperScan(value = "com.heeverse.*.domain.mapper", annotationClass = LockScanMapper.class,
-        sqlSessionFactoryRef = "primarySqlSessionFactory")
-public class SqlSessionFactoryConfiguration {
-
-    @Primary
-    @Bean(name = "primarySqlSessionFactory")
-    public SqlSessionFactory primarySqlSessionFactory(@Qualifier("primaryDataSource") DataSource primaryDataSource, ApplicationContext applicationContext) throws Exception {
-        SqlSessionFactoryBean sqlSessionFactoryBean = new SqlSessionFactoryBean();
-        sqlSessionFactoryBean.setDataSource(primaryDataSource);
-        sqlSessionFactoryBean.setConfigLocation(new ClassPathResource("mybatis-config.xml"));
-        sqlSessionFactoryBean.setMapperLocations(applicationContext.getResources("classpath:mapper/basic/**.xml"));
-        return sqlSessionFactoryBean.getObject();
-    }
-
-
-    @Bean(name = "primarySqlSessionTemplate")
-    public SqlSessionTemplate primarySqlSessionTemplate(@Qualifier("primarySqlSessionFactory") SqlSessionFactory primarySqlSessionFactory) {
-        return new SqlSessionTemplate(primarySqlSessionFactory);
-    }
-
-}
