+++ src/main/resources/application-dev.yml
@@ -1,9 +1,16 @@
 spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
   sql:
     init:
       mode: never
-  jmx:
-    enabled: false
+
+
+mybatis:
+  mapper-locations: classpath:mapper/*.xml
+  configuration:
+    map-underscore-to-camel-case: true
+    arg-name-based-constructor-auto-mapping: true
 
 logging:
   level:
