+++ src/main/resources/application-local.yml
@@ -13,9 +13,7 @@ spring:
       mode: embedded
 
 mybatis:
-  mapper-locations:
-    - classpath:mapper/basic/*.xml
-    - classpath:mapper/lock-mapper/*.xml
+  mapper-locations: classpath:mapper/*.xml
   configuration:
     map-underscore-to-camel-case: true
     arg-name-based-constructor-auto-mapping: true
