+++ src/main/resources/application-prod.yml
@@ -1,6 +1,18 @@
 spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    hikari:
+      register-mbeans: true
   jmx:
     enabled: false
+
+
+
+mybatis:
+  mapper-locations: classpath:mapper/*.xml
+  configuration:
+    map-underscore-to-camel-case: true
+    arg-name-based-constructor-auto-mapping: true
 
 logging:
   level:
